+++ src/unfold/contrib/inlines/admin.py
@@ -1,21 +1,18 @@
 from functools import partial
 from typing import Any
 
-from django import forms
+from django.contrib.admin.options import InlineModelAdmin
 from django.contrib.admin.utils import NestedObjects, flatten_fieldsets
 from django.core.exceptions import ValidationError
 from django.db import router
-from django.db.models import Model
+from django.db.models import Model, QuerySet
 from django.forms.formsets import DELETION_FIELD_NAME
-from django.forms.models import modelform_defines_fields
+from django.forms.models import ALL_FIELDS, modelform_defines_fields
 from django.http import HttpRequest
 from django.utils.text import get_text_list
 from django.utils.translation import gettext_lazy as _
 
-from unfold.admin import (
-    StackedInline,
-    TabularInline,
-)
+from unfold.admin import StackedInline, TabularInline
 from unfold.contrib.inlines.checks import NonrelatedModelAdminChecks
 from unfold.contrib.inlines.forms import (
     NonrelatedInlineModelFormSet,
@@ -23,7 +20,7 @@ from unfold.contrib.inlines.forms import (
 )
 
 
-class NonrelatedInlineMixin:
+class NonrelatedInlineMixin(InlineModelAdmin):
     checks_class = NonrelatedModelAdminChecks
     formset = NonrelatedInlineModelFormSet
 
@@ -39,6 +36,12 @@ class NonrelatedInlineMixin:
         return nonrelated_inline_formset_factory(
             self.model, save_new_instance=self.save_new_instance, **defaults
         )
+
+    def get_form_queryset(self, obj: Model) -> QuerySet:
+        raise NotImplementedError("get_form_queryset must be implemented")
+
+    def save_new_instance(self, parent: Model, instance: Model) -> None:
+        raise NotImplementedError("save_new_instance must be implemented")
 
     def _get_formset_defaults(
         self, request: HttpRequest, obj: Model | None = None, **kwargs: Any
@@ -89,7 +92,7 @@ class NonrelatedInlineMixin:
                     collector = NestedObjects(using=using)
                     if self.instance._state.adding:
                         return
-                    collector.collect([self.instance])
+                    collector.collect([self.instance])  # type: ignore
                     if collector.protected:
                         objs = []
                         for p in collector.protected:
@@ -134,7 +137,7 @@ class NonrelatedInlineMixin:
         if defaults["fields"] is None and not modelform_defines_fields(
             defaults["form"]
         ):
-            defaults["fields"] = forms.ALL_FIELDS
+            defaults["fields"] = ALL_FIELDS
 
         return defaults